+++ packages/datatrak-web/src/utils/index.ts
@@ -5,4 +5,3 @@
 
 export { errorToast, successToast } from './toast';
 export { useDebounce } from './useDebounce';
-export { useUserCountries } from './useUserCountries';
