+++ packages/datatrak-web/src/views/SurveySelectPage/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { SurveySelectPage } from './SurveySelectPage';
